refactor(auth): group OAuth feature
