+++ app/Features/OAuth/Helpers/OidcLogoutTokenValidator.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Services\OAuth;
+namespace App\Features\OAuth\Helpers;
 
 use App\Models\OAuthConnection;
 use Firebase\JWT\JWK;
