+++ app/Filament/Resources/OAuthConnections/Pages/ManageOAuthConnections.php
@@ -4,12 +4,12 @@ declare(strict_types=1);
 
 namespace App\Filament\Resources\OAuthConnections\Pages;
 
+use App\Features\OAuth\Helpers\OAuthConnectionService;
+use App\Features\OAuth\Helpers\SocialiteProviderRegistry;
 use App\Filament\Actions\Drawer;
 use App\Filament\Resources\OAuthConnections\OAuthConnectionResource;
 use App\Filament\Resources\OAuthConnections\Schemas\OAuthConnectionForm;
 use App\Models\OAuthConnection;
-use App\Services\OAuth\OAuthConnectionService;
-use App\Services\OAuth\SocialiteProviderRegistry;
 use Filafly\Icons\Phosphor\Enums\Phosphor;
 use Filament\Actions\CreateAction;
 use Filament\Resources\Pages\ManageRecords;
