+++ app/Http/Controllers/Auth/RegisterController.php
@@ -4,18 +4,18 @@ namespace App\Http\Controllers\Auth;
 
 use App\Contracts\Controller;
 use App\Enums\UserState;
+use App\Features\OAuth\Helpers\ExternalAuthSessionService;
+use App\Features\OAuth\Helpers\OAuthConnectionService;
+use App\Features\OAuth\Helpers\OAuthIdentityService;
+use App\Features\OAuth\Helpers\OAuthSessionStore;
 use App\Models\Airline;
 use App\Models\Invite;
 use App\Models\User;
 use App\Models\UserField;
 use App\Models\UserFieldValue;
-use App\Services\OAuth\ExternalAuthSessionService;
-use App\Services\OAuth\OAuthConnectionService;
-use App\Services\OAuth\OAuthIdentityService;
 use App\Services\UserService;
 use App\Support\Countries;
 use App\Support\HttpClient;
-use App\Support\OAuthSessionStore;
 use App\Support\Timezonelist;
 use Exception;
 use Illuminate\Contracts\Validation\Validator;
