+++ app/Http/Controllers/Frontend/ProfileController.php
@@ -5,6 +5,8 @@ namespace App\Http\Controllers\Frontend;
 use App\Addons\AddonRegistry;
 use App\Contracts\Controller;
 use App\Events\ProfileUpdated;
+use App\Features\OAuth\Helpers\OAuthConnectionService;
+use App\Features\OAuth\Helpers\SocialiteProviderRegistry;
 use App\Features\Tour\Enums\TourStatus;
 use App\Http\Data\ProfileData;
 use App\Models\Airline;
@@ -12,8 +14,6 @@ use App\Models\Award;
 use App\Models\User;
 use App\Models\UserField;
 use App\Models\UserFieldValue;
-use App\Services\OAuth\OAuthConnectionService;
-use App\Services\OAuth\SocialiteProviderRegistry;
 use App\Services\UserService;
 use App\Support\ApiScope;
 use App\Support\Countries;
