+++ app/Providers/AppServiceProvider.php
@@ -12,6 +12,8 @@ use App\Enums\PirepState;
 use App\Enums\UserState;
 use App\Features\Assets\AssetService;
 use App\Features\Assets\AssetTypes;
+use App\Features\OAuth\Helpers\OAuthConnectionService;
+use App\Features\OAuth\Helpers\SocialiteProviderRegistry;
 use App\Http\Composers\PageLinksComposer;
 use App\Http\Composers\VersionComposer;
 use App\Models\Role;
@@ -20,8 +22,6 @@ use App\Policies\Filament\ActivityPolicy;
 use App\Services\AddonSettingService;
 use App\Services\ImageUploadService;
 use App\Services\ModuleService;
-use App\Services\OAuth\OAuthConnectionService;
-use App\Services\OAuth\SocialiteProviderRegistry;
 use App\Services\PermissionRegistry;
 use App\Services\RouteForge\Contracts\LintRule;
 use App\Services\RouteForge\LintRunner;
