+++ tests/Feature/OAuthTest.php
@@ -2,6 +2,9 @@
 
 declare(strict_types=1);
 
+use App\Features\OAuth\Helpers\ExternalAuthSessionService;
+use App\Features\OAuth\Helpers\OAuthConnectionService;
+use App\Features\OAuth\Helpers\SocialiteProviderRegistry;
 use App\Http\Controllers\Auth\OAuthController;
 use App\Http\Controllers\Auth\RegisterController;
 use App\Models\Airline;
@@ -12,9 +15,6 @@ use App\Models\OAuthConnection;
 use App\Models\User;
 use App\Models\UserIdentity;
 use App\Models\UserOAuthToken;
-use App\Services\OAuth\ExternalAuthSessionService;
-use App\Services\OAuth\OAuthConnectionService;
-use App\Services\OAuth\SocialiteProviderRegistry;
 use Illuminate\Auth\Events\Login as AuthLogin;
 use Illuminate\Contracts\Http\Kernel as HttpKernel;
 use Illuminate\Http\RedirectResponse;
